Email promises & Verifing
Promises are now used within the email class to allow better control over the flow of emails by returning promises that can be caught and handled. 

Verifing is now created to allow the server to know the status of the connection for stopping email flow if required by using the nodemailer verify function.