Endspoints & Contact endpoint
Contact us endpoint to allow recieving contact us requests from the client to then be sent to the contactUs email by listening to post requests on '/contact/send' 

Status can be requested by 'api/contact/status' this status will be a 200 or a 500 http code based on if the email service is currenlty live. 

Each email request is validated and checked before sending. DenyInvalidAndBlockedDomains is currently skipped.