Fixed travis, preplanning code and commenting existing code.
Moved travis to the root folder and now the install process moves into the project folder, this allows the travis file to be found and builds to be made.

Fixed a http status code from 500 to 503 where the email service would be unavailable

Fixed contactEndpoint by correctly contacting requiring the correct parms and removed the param requirements within status.

Introduced the volunteerEndpoint and editorEndpoint, preplanning code for hooking into the volunteer web and updating content with the page editor. 

commented existing code.